# bcsc-api/src/gen/java/io/swagger/model/LotIngredientItem.java
/*
 * Swagger PetNutrition YAML
 * Pet Nutrition
 *
 * OpenAPI spec version: 1.0.0
 * Contact: pzybrick@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package io.swagger.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import io.swagger.model.LotSupplierBlockItem;
import java.util.ArrayList;
import java.util.List;
import javax.validation.constraints.*;

/**
 * LotIngredientItem
 */
@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaJerseyServerCodegen", date = "2018-04-20T00:28:28.441Z")
public class LotIngredientItem   {
  @JsonProperty("ingredientSequence")
  private Integer ingredientSequence = null;

  @JsonProperty("ingredientName")
  private String ingredientName = null;

  @JsonProperty("lotSupplierBlockItems")
  private List<LotSupplierBlockItem> lotSupplierBlockItems = null;

  public LotIngredientItem ingredientSequence(Integer ingredientSequence) {
    this.ingredientSequence = ingredientSequence;
    return this;
  }

  /**
   * Get ingredientSequence
   * @return ingredientSequence
   **/
  @JsonProperty("ingredientSequence")
  @ApiModelProperty(example = "12", value = "")
  public Integer getIngredientSequence() {
    return ingredientSequence;
  }

  public void setIngredientSequence(Integer ingredientSequence) {
    this.ingredientSequence = ingredientSequence;
  }

  public LotIngredientItem ingredientName(String ingredientName) {
    this.ingredientName = ingredientName;
    return this;
  }

  /**
   * Get ingredientName
   * @return ingredientName
   **/
  @JsonProperty("ingredientName")
  @ApiModelProperty(example = "Brown Rice", value = "")
  public String getIngredientName() {
    return ingredientName;
  }

  public void setIngredientName(String ingredientName) {
    this.ingredientName = ingredientName;
  }

  public LotIngredientItem lotSupplierBlockItems(List<LotSupplierBlockItem> lotSupplierBlockItems) {
    this.lotSupplierBlockItems = lotSupplierBlockItems;
    return this;
  }

  public LotIngredientItem addLotSupplierBlockItemsItem(LotSupplierBlockItem lotSupplierBlockItemsItem) {
    if (this.lotSupplierBlockItems == null) {
      this.lotSupplierBlockItems = new ArrayList<LotSupplierBlockItem>();
    }
    this.lotSupplierBlockItems.add(lotSupplierBlockItemsItem);
    return this;
  }

  /**
   * Get lotSupplierBlockItems
   * @return lotSupplierBlockItems
   **/
  @JsonProperty("lotSupplierBlockItems")
  @ApiModelProperty(value = "")
  public List<LotSupplierBlockItem> getLotSupplierBlockItems() {
    return lotSupplierBlockItems;
  }

  public void setLotSupplierBlockItems(List<LotSupplierBlockItem> lotSupplierBlockItems) {
    this.lotSupplierBlockItems = lotSupplierBlockItems;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    LotIngredientItem lotIngredientItem = (LotIngredientItem) o;
    return Objects.equals(this.ingredientSequence, lotIngredientItem.ingredientSequence) &&
        Objects.equals(this.ingredientName, lotIngredientItem.ingredientName) &&
        Objects.equals(this.lotSupplierBlockItems, lotIngredientItem.lotSupplierBlockItems);
  }

  @Override
  public int hashCode() {
    return Objects.hash(ingredientSequence, ingredientName, lotSupplierBlockItems);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class LotIngredientItem {\n");
    
    sb.append("    ingredientSequence: ").append(toIndentedString(ingredientSequence)).append("\n");
    sb.append("    ingredientName: ").append(toIndentedString(ingredientName)).append("\n");
    sb.append("    lotSupplierBlockItems: ").append(toIndentedString(lotSupplierBlockItems)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}