# bcsc-api/src/gen/java/io/swagger/model/LotTreeItem.java
/*
 * Swagger PetNutrition YAML
 * Pet Nutrition
 *
 * OpenAPI spec version: 1.0.0
 * Contact: pzybrick@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package io.swagger.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import io.swagger.model.LotIngredientItem;
import java.util.ArrayList;
import java.util.List;
import javax.validation.constraints.*;

/**
 * LotTreeItem
 */
@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaJerseyServerCodegen", date = "2018-04-20T00:28:28.441Z")
public class LotTreeItem   {
  @JsonProperty("manufacturerLotNumber")
  private String manufacturerLotNumber = null;

  @JsonProperty("manufacturerLotFilledDate")
  private String manufacturerLotFilledDate = null;

  @JsonProperty("lotIngredientItems")
  private List<LotIngredientItem> lotIngredientItems = null;

  public LotTreeItem manufacturerLotNumber(String manufacturerLotNumber) {
    this.manufacturerLotNumber = manufacturerLotNumber;
    return this;
  }

  /**
   * Get manufacturerLotNumber
   * @return manufacturerLotNumber
   **/
  @JsonProperty("manufacturerLotNumber")
  @ApiModelProperty(example = "LOT-20180102", value = "")
  public String getManufacturerLotNumber() {
    return manufacturerLotNumber;
  }

  public void setManufacturerLotNumber(String manufacturerLotNumber) {
    this.manufacturerLotNumber = manufacturerLotNumber;
  }

  public LotTreeItem manufacturerLotFilledDate(String manufacturerLotFilledDate) {
    this.manufacturerLotFilledDate = manufacturerLotFilledDate;
    return this;
  }

  /**
   * Get manufacturerLotFilledDate
   * @return manufacturerLotFilledDate
   **/
  @JsonProperty("manufacturerLotFilledDate")
  @ApiModelProperty(example = "", value = "")
  public String getManufacturerLotFilledDate() {
    return manufacturerLotFilledDate;
  }

  public void setManufacturerLotFilledDate(String manufacturerLotFilledDate) {
    this.manufacturerLotFilledDate = manufacturerLotFilledDate;
  }

  public LotTreeItem lotIngredientItems(List<LotIngredientItem> lotIngredientItems) {
    this.lotIngredientItems = lotIngredientItems;
    return this;
  }

  public LotTreeItem addLotIngredientItemsItem(LotIngredientItem lotIngredientItemsItem) {
    if (this.lotIngredientItems == null) {
      this.lotIngredientItems = new ArrayList<LotIngredientItem>();
    }
    this.lotIngredientItems.add(lotIngredientItemsItem);
    return this;
  }

  /**
   * Get lotIngredientItems
   * @return lotIngredientItems
   **/
  @JsonProperty("lotIngredientItems")
  @ApiModelProperty(value = "")
  public List<LotIngredientItem> getLotIngredientItems() {
    return lotIngredientItems;
  }

  public void setLotIngredientItems(List<LotIngredientItem> lotIngredientItems) {
    this.lotIngredientItems = lotIngredientItems;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    LotTreeItem lotTreeItem = (LotTreeItem) o;
    return Objects.equals(this.manufacturerLotNumber, lotTreeItem.manufacturerLotNumber) &&
        Objects.equals(this.manufacturerLotFilledDate, lotTreeItem.manufacturerLotFilledDate) &&
        Objects.equals(this.lotIngredientItems, lotTreeItem.lotIngredientItems);
  }

  @Override
  public int hashCode() {
    return Objects.hash(manufacturerLotNumber, manufacturerLotFilledDate, lotIngredientItems);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class LotTreeItem {\n");
    
    sb.append("    manufacturerLotNumber: ").append(toIndentedString(manufacturerLotNumber)).append("\n");
    sb.append("    manufacturerLotFilledDate: ").append(toIndentedString(manufacturerLotFilledDate)).append("\n");
    sb.append("    lotIngredientItems: ").append(toIndentedString(lotIngredientItems)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}